requirements.txt
# Google Agent Development Kit
google-adk

# For environment variable management (optional but recommended)
python-dotenv

